+++ src/librustc_session/Cargo.toml
@@ -14,3 +14,4 @@ rustc_errors = { path = "../librustc_errors" }
 rustc_serialize = { path = "../libserialize", package = "serialize" }
 rustc_data_structures = { path = "../librustc_data_structures" }
 syntax_pos = { path = "../libsyntax_pos" }
+rustc_index = { path = "../librustc_index" }
