+++ src/librustc_session/lib.rs
@@ -2,3 +2,4 @@ pub mod cgu_reuse_tracker;
 pub mod utils;
 #[macro_use]
 pub mod lint;
+pub mod node_id;
