+++ src/libsyntax/early_buffered_lints.rs
@@ -3,8 +3,6 @@
 //! Since we cannot have a dependency on `librustc`, we implement some types here that are somewhat
 //! redundant. Later, these types can be converted to types for use by the rest of the compiler.
 
-use crate::ast::NodeId;
-use syntax_pos::MultiSpan;
 use rustc_session::lint::FutureIncompatibleInfo;
 use rustc_session::declare_lint;
 pub use rustc_session::lint::BufferedEarlyLint;
@@ -29,19 +27,4 @@ declare_lint! {
     pub INCOMPLETE_INCLUDE,
     Deny,
     "trailing content in included file"
-}
-
-/// Stores buffered lint info which can later be passed to `librustc`.
-pub struct BufferedEarlyLint {
-    /// The span of code that we are linting on.
-   pub span: MultiSpan,
-
-   /// The lint message.
-   pub msg: String,
-
-   /// The `NodeId` of the AST node that generated the lint.
-   pub id: NodeId,
-
-   /// A lint Id that can be passed to `rustc::lint::Lint::from_parser_lint_id`.
-   pub lint_id: &'static rustc_session::lint::Lint,
 }
